Move new decorators to new namespace
