+++ src/FormDecoration/Transformation.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ipl\Html\FormDecorator;
+namespace ipl\Html\FormDecoration;
 
 /**
  * Describes how the content should be transformed
