+++ tests/FormDecorator/DecoratorChainTest.php
@@ -4,7 +4,7 @@ namespace ipl\Tests\Html\FormDecorator;
 
 use Exception;
 use InvalidArgumentException;
-use ipl\Html\FormDecorator\DecoratorChain;
+use ipl\Html\FormDecoration\DecoratorChain;
 use ipl\Html\FormElement\TextElement;
 use ipl\Tests\Html\TestCase;
 
@@ -190,7 +190,7 @@ class DecoratorChainTest extends TestCase
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage(
             'Expects array value at position 0 to be a string or an instance of ipl\Html\Contract\Decorator,'
-            . ' got ipl\Html\FormDecorator\DecoratorChain instead'
+            . ' got ipl\Html\FormDecoration\DecoratorChain instead'
         );
 
         $this->chain->addDecorators([$this->chain]);
