+++ tests/FormDecorator/TestSkipRenderElementDecorator.php
@@ -4,7 +4,7 @@ namespace ipl\Tests\Html\FormDecorator;
 
 use ipl\Html\Contract\Decorator;
 use ipl\Html\Contract\FormElement;
-use ipl\Html\FormDecorator\DecorationResults;
+use ipl\Html\FormDecoration\DecorationResults;
 
 /**
  * Render the $formElement and skip "TestRenderElement" decorator
